Squashed commit of the following:
commit bda149459fc7d88d237bd6ecd083894dec41595f
Author: Benjamin Hindman <benh@twitter.com>
Date:   Fri Jun 3 14:28:42 2011 -0700

    Updates from jsirois and wfarner code review.

commit a7c5b86037af95f618bbf35ee330cb4a774cbf85
Author: Benjamin Hindman <benh@twitter.com>
Date:   Fri Jun 3 14:28:22 2011 -0700

    Getting LXC up to date, still needs a compile and test on Linux.

commit b9aaf1cfba809fe25fb0caea8e374612e7c5b950
Merge: 0335bcd 27b8b30
Author: Benjamin Hindman <benh@twitter.com>
Date:   Fri Jun 3 12:29:23 2011 -0700

    Merge branch 'master' into executor-shutdown

commit 0335bcd5092179b94c525da52fac870fa31c49c2
Author: Benjamin Hindman <benh@twitter.com>
Date:   Thu Jun 2 17:21:38 2011 -0700

    Re-architected how executor shutdowns work so that an executor can gracefully shutdown within a configurable timeout.

+++ src/slave/process_based_isolation_module.cpp
@@ -6,6 +6,7 @@
 
 #include "process_based_isolation_module.hpp"
 
+#include "common/foreach.hpp"
 #include "common/type_utils.hpp"
 
 using namespace mesos;
@@ -21,7 +22,11 @@ using std::string;
 
 
 ProcessBasedIsolationModule::ProcessBasedIsolationModule()
+  : initialized(false)
 {
+  // Spawn the reaper, note that it might send us a message before we
+  // actually get spawned ourselves, but that's okay, the message will
+  // just get dropped.
   reaper = new Reaper();
   spawn(reaper);
   dispatch(reaper, &Reaper::addProcessExitedListener, this);
@@ -60,12 +65,13 @@ void ProcessBasedIsolationModule::launchExecutor(
     LOG(FATAL) << "Cannot launch executors before initialization!";
   }
 
-  LOG(INFO) << "Starting executor '" << executorInfo.uri()
-            << "' for framework " << frameworkId;
+  LOG(INFO) << "Launching '" << executorInfo.uri()
+            << "' for executor '" << executorInfo.executor_id()
+            << "' of framework " << frameworkId;
 
   pid_t pid;
   if ((pid = fork()) == -1) {
-    PLOG(ERROR) << "Failed to fork to launch new executor";
+    PLOG(FATAL) << "Failed to fork to launch new executor";
   }
 
   if (pid) {
@@ -95,16 +101,28 @@ void ProcessBasedIsolationModule::killExecutor(
     const FrameworkID& frameworkId,
     const ExecutorID& executorId)
 {
+  if (!pgids.contains(frameworkId) ||
+      !pgids[frameworkId].contains(executorId)) {
+    LOG(ERROR) << "ERROR! Asked to kill an unknown executor!";
+    return;
+  }
+
   if (pgids[frameworkId][executorId] != -1) {
     // TODO(benh): Consider sending a SIGTERM, then after so much time
     // if it still hasn't exited do a SIGKILL (can use a libprocess
-    // process for this).
+    // process for this). This might not be necessary because we have
+    // higher-level semantics via the first shut down phase that gets
+    // initiated by the slave.
     LOG(INFO) << "Sending SIGKILL to process group "
               << pgids[frameworkId][executorId];
 
     killpg(pgids[frameworkId][executorId], SIGKILL);
 
     pgids[frameworkId].erase(executorId);
+
+    if (pgids[frameworkId].size() == 0) {
+      pgids.erase(frameworkId);
+    }
 
     // TODO(benh): Kill all of the process's descendants? Perhaps
     // create a new libprocess process that continually tries to kill
